+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Develop
 
 - Added option to remove wdg from the list
+- Added option to set watchdog name and print its name on error
 
 ## v0.0.1